fix warning

+++ libgo/common/anys.h
@@ -181,7 +181,7 @@ public:
     inline void* align(size_t __align, size_t __size, void*& __ptr, size_t& __space) noexcept
     {
         const auto __intptr = reinterpret_cast<uintptr_t>(__ptr);
-        const auto __aligned = (__intptr - 1u + __align) & -__align;
+        const auto __aligned = (__intptr - 1u + __align) & -(int)__align;
         const auto __diff = __aligned - __intptr;
         if ((__size + __diff) > __space)
             return nullptr;
